requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
PyPDF2==3.0.1
email-validator==2.1.0
huggingface_hub==0.24.2




